Moves thumbnails.js to the gallery directory
