+++ js/gallery/thumbnails.js
