+++ js/main.js
@@ -1,4 +1,4 @@
 import { createPhotosData } from './data.js';
-import { renderThumbnails } from './thumbnails.js';
+import { renderThumbnails } from './gallery/thumbnails.js';
 
 renderThumbnails(createPhotosData());
